% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central_veins.R
\name{central_veins}
\alias{central_veins}
\title{Central Vein Detection}
\usage{
central_veins(
  epi,
  t1,
  flair,
  prob_map = NULL,
  bin_map = NULL,
  verbose = FALSE,
  bias_correct = TRUE,
  register = TRUE,
  skull_strip = TRUE
)
}
\arguments{
\item{epi}{T2*-EPI volume of class \code{antsImage}, class \code{nifti}, or path to ".nii.gz" file.}

\item{t1}{T1-weighted volume of class \code{antsImage}, class \code{nifti}, or path to ".nii.gz" file.}

\item{flair}{T2-FLAIR volume of class \code{antsImage}, class \code{nifti}, or path to ".nii.gz" file.}

\item{prob_map}{volume of class \code{antsImage}, class \code{nifti},  or path to ".nii.gz" file containing the probability that each voxel is a lesion voxel.
If a probability map is not included, the MIMoSA model will be applied (Valcarcel et al., 2017).}

\item{bin_map}{binarized volume of class\code{antsImage}, class \code{nifti}, or path to ".nii.gz" file. Mask in which voxels are classified as either lesion voxels or not lesion voxels.}

\item{verbose}{logical value for printing pipeline updates (default = FALSE)}

\item{bias_correct}{a logical value reflecting whether bias correction still needs to be performed (default = TRUE)}

\item{register}{a logical value reflecting whether images still need to be registered (default = TRUE)}

\item{skull_strip}{a logical value reflecting whether skull stripping still needs to be performed (default = TRUE)}
}
\value{
A list containing candidate.lesions (a nifti file with labeled lesions evaluated for CVS),
cvs.prob_map (a nifti file in which candidate lesions are labeled with their CVS probability), and
cvs.biomarker (a numeric value representing the average CVS probability of a subject's lesions).
}
\description{
This function obtains the probability that each lesion in a subject's deep white-matter has a central vein.
}
\examples{
\dontrun{
library(ANTsRCore)
epi <- check_ants("path/to/epi")
flair <- check_ants("path/to/flair")
t1 <- check_ants("path/to/t1")
cvs <- centralveins(epi = epi, t1 = t1, flair = flair)
}
}
