% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtboundary.R
\name{dtboundary}
\alias{dtboundary}
\title{Distance to Mask Boundary}
\usage{
dtboundary(mask)
}
\arguments{
\item{mask}{a 3D array or image of class \code{nifti}, containing a binary mask where 1 represents structure.}
}
\value{
A new image in which voxels have been assigned their distance to the nearest boundary.
}
\description{
This function finds the distance of each voxel to the nearest boundary in a given mask.
}
\examples{
\dontrun{
library(neurobase)
lesion.mask <- readnii('path/to/mask')
dtb <- dtboundary(mask = lesion.mask) }
}
