% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelreg.R
\name{labelreg}
\alias{labelreg}
\title{Register Label Image Based on Full Image}
\usage{
labelreg(fullimage, labelimage, fixedimage, typeofTransform = "Rigid",
  interpolator = "lanczosWindowedSinc")
}
\arguments{
\item{fullimage}{an image of class \code{nifti}, in the same space as the label image, which will be registered to the fixed image.}

\item{labelimage}{an image of class \code{nifti} with limited structural information, in the same space as the full image, to which the full image registration will be applied.}

\item{fixedimage}{an image of class \code{nifti}, to which the other images will be registered.}

\item{typeofTransform}{the type of registration desired; this value is passed onto the registration function from extrantsr.}

\item{interpolator}{the type of interpolation desired; this value is passed onto the antsApplyTransforms function from ANTsRCore}
}
\value{
A list containing image_reg (the registered version of fullimage) and label_reg (the registered version of labelimage).
}
\description{
This function registers a full image to a fixed image, then applies the registration to a label or binary image in the same space as the full image.
}
\examples{
\dontrun{
library(neurobase)
library(fslr)
flair <- readnii('path/to/flair')
t1 <- readnii('path/to/t1')
tissue.class= <- fast(t1, opts='--nobias')
registered <- labelreg(t1,tissue.class,flair)
t1_reg = registered$image_reg
tissue.class_reg = registered$label_reg }
}
