% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesioncenters.R
\name{lesioncenters}
\alias{lesioncenters}
\title{Distinct Lesion Centers}
\usage{
lesioncenters(probmap, binmap, minCenterSize = 10, radius = 1,
  parallel = F, cores = 2)
}
\arguments{
\item{probmap}{a 3D array or image of class \code{nifti}, containing the probability that each voxel is a lesion voxel}

\item{binmap}{a 3D array or \code{nifti} mask in which voxels are classified as either lesion voxels or not lesion voxels.
Note that mask should be in the same space as the probmap volume}

\item{minCenterSize}{an integer value representing the minimum number of connected voxels that can be considered a lesion center}

\item{radius}{an integer specifying radius of the neighborhood (in voxels) for which the hessian should be calculated.}

\item{parallel}{is a logical value that indicates whether the user's computer
is Linux or Unix (i.e. macOS), and should run the code in parallel}

\item{cores}{if parallel = TRUE, cores is an integer value that indicates how many cores
the function should be run on}
}
\value{
A list containing lesioncenters (a nifti file with labeled lesion centers) and lesioncount (an integer value representing the number of distinct lesions)
}
\description{
This function finds the centers of distinct lesions based on a lesion probability map. The method is described in Dworkin et al., (2018).
}
\examples{
\dontrun{
library(neurobase)
lesion.probs <- readnii('path/to/probabilitymap')
centers <- lesioncenters(probmap = lesion.probs, binmap = lesion.probs>0.30,
                         parallel = TRUE, cores = 4) }
}
\references{
J.D. Dworkin, K.A. Linn, I. Oguz, G.M. Fleishman, R. Bakshi, G. Nair, P.A. Calabresi, R.G. Henry, J. Oh, N. Papinutto, D. Pelletier, W. Rooney, W. Stern, N.L. Sicotte, D.S. Reich, R.T. Shinohara. An automated statistical technique for counting distinct multiple sclerosis lesions. American Journal of Neuroradiology, 2018; 39, 626-633.
}
